More environment variable specific API
