+++ core/jvm-native/src/main/scala/zio/SystemPlatformSpecific.scala
@@ -25,40 +25,13 @@ import scala.collection.JavaConverters._
 
 private[zio] trait SystemPlatformSpecific { self: System.type =>
 
-  trait SystemLivePlatformSpecific extends System {
-    @transient override val unsafe: UnsafeAPI =
-      new UnsafeAPI {
-        override def env(variable: String)(implicit unsafe: Unsafe): Option[String] =
-          Option(JSystem.getenv(variable))
+  private[zio] val environmentProvider = new EnvironmentProvider {
+    override def env(variable: String): Option[String] =
+      Option(JSystem.getenv(variable))
 
-        override def envOrElse(variable: String, alt: => String)(implicit unsafe: Unsafe): String =
-          envOrElseWith(variable, alt)(env)
-
-        override def envOrOption(variable: String, alt: => Option[String])(implicit unsafe: Unsafe): Option[String] =
-          envOrOptionWith(variable, alt)(env)
-
-        @nowarn("msg=JavaConverters")
-        override def envs()(implicit unsafe: Unsafe): Map[String, String] =
-          JSystem.getenv.asScala.toMap
-
-        override def lineSeparator()(implicit unsafe: Unsafe): String =
-          JSystem.lineSeparator
-
-        @nowarn("msg=JavaConverters")
-        override def properties()(implicit unsafe: Unsafe): Map[String, String] =
-          JSystem.getProperties.asScala.toMap
-
-        override def property(prop: String)(implicit unsafe: Unsafe): Option[String] =
-          Option(JSystem.getProperty(prop))
-
-        override def propertyOrElse(prop: String, alt: => String)(implicit unsafe: Unsafe): String =
-          propertyOrElseWith(prop, alt)(property)
-
-        override def propertyOrOption(prop: String, alt: => Option[String])(implicit
-          unsafe: Unsafe
-        ): Option[String] =
-          propertyOrOptionWith(prop, alt)(property)
-      }
+    @nowarn("msg=JavaConverters")
+    override def envs: Map[String, String] =
+      JSystem.getenv().asScala.toMap
   }
 
 }
